requirements.txt
flask==2.1.0
selenium==4.4.0
google-generativeai==0.3.0
flask-limiter==3.3.0
flask-caching==2.0.2
python-dotenv==1.0.0
webdriver-manager==3.8.2
streamlit==1.32.0
pandas==2.2.0
plotly==5.18.0

